requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
Levenshtein==0.23.0
multidict==6.0.4
numpy==1.26.3
pandas==2.1.4
phonenumbers==8.13.25
psycopg2-binary==2.9.9
pycountry==22.3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.23.0
pytz==2023.3.post1
rapidfuzz==3.5.2
requests==2.31.0
sellix-python-sdk==1.0.9
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
stubs==1.0.0
tzdata==2023.4
urllib3==2.1.0
uvloop==0.19.0
yarl==1.9.2
